Fixed some game and dealer issues
Fixed the blackjack result that is now handled correctly. Also fixed the card draw and deck system, in fact before it was not handled well and when the cards ran out it gave an error and the system blocked. Fixed bet placement (now you can't play if you don't place a bet first) and if your balance is insufficient to place the bet a message is displayed.